requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
google-genai==1.63.0
pdfplumber==0.11.9
python-docx==1.2.0
chromadb==1.5.0
langchain==1.2.10
langchain-core==1.2.12
langchain-google-genai==4.2.0
langchain-chroma==1.1.0
langchain-text-splitters==1.1.0
